requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
streamlit>=1.28.0
pillow>=10.0.0
numpy>=1.24.0
safetensors>=0.3.0
xformers>=0.0.20; sys_platform != "darwin"
